refactor: PointBalance VO 패키지 이동 반영
